Update system install guide
Update the system install guide with changes that I had to make when I
set up my system back in September.

Remove the reference to `NODE_PATH` because it seems like that's still
supported but not encouraged. I didn't set it when I set up my system
and haven't encountered any problems.

Add a mention of Atom in the editors section because it seems like a lot
of people use it, and it's my preferred non-vim editor.